requirements.txt
Flask
python-dotenv
requests
openai
beautifulsoup4
python-dateutil
jsonschema
Pillow
duckdb
tiktoken
chromadb
transformers
sentence-transformers
uvloop
fastapi
uvicorn
httpx
markdown2
whisper